python-openai-nlp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai>=1.12.0
python-jose==3.3.0
requests==2.31.0 